EES-3369 - completed public table builder performance test.  Renamed dataService to adminService in place of new dataService to represent Data API.  README correction on firewall rules
