+++ tests/performance-tests/src/utils/dataService.ts
@@ -1,21 +1,8 @@
-/* eslint-disable no-console */
-import { sleep } from 'k6';
-import http, { RefinedResponse } from 'k6/http';
 import HttpClient from './httpClient';
-import TestData from '../tests/testData';
 
 const applicationJsonHeaders = {
   'Content-Type': 'application/json',
 };
-
-type OverallStage =
-  | 'Validating'
-  | 'Complete'
-  | 'Failed'
-  | 'Invalid'
-  | 'Scheduled'
-  | 'Started'
-  | 'Superseded';
 
 export interface SubjectMeta {
   filters: {
@@ -51,404 +38,14 @@ export interface SubjectMeta {
   };
 }
 
-interface Topic {
-  id: string;
-  title: string;
-  themeId: string;
-}
-
-interface ThemeAndTopics {
-  id: string;
-  title: string;
-  topics: Topic[];
-}
-
-interface Release {
-  id: string;
-  year: number;
-}
-
-interface PublicationAndReleases {
-  id: string;
-  topicId: string;
-  title: string;
-  releases: Release[];
-}
-
 class DataService {
   private readonly client: HttpClient;
 
-  constructor(
-    adminUrl: string,
-    accessToken: string,
-    checkResponseStatus = true,
-  ) {
+  constructor(dataApiUrl: string, checkResponseStatus = true) {
     this.client = new HttpClient({
-      baseUrl: adminUrl,
-      accessToken,
+      baseUrl: dataApiUrl,
       checkResponseStatus,
     });
-  }
-
-  getThemes() {
-    const { json } = this.client.get<ThemeAndTopics[]>(
-      `/api/themes`,
-      applicationJsonHeaders,
-    );
-    return json;
-  }
-
-  getTheme({ title }: { title: string }): ThemeAndTopics | undefined {
-    return this.getThemes().find(
-      ({ title: existingTitle }) => existingTitle === title,
-    );
-  }
-
-  createTheme({ title }: { title: string }) {
-    const { response, json } = this.client.post<{ id: string }>(
-      `/api/themes`,
-      JSON.stringify({
-        title,
-      }),
-      applicationJsonHeaders,
-    );
-
-    console.log(`Created Theme ${title}`);
-
-    return {
-      id: json.id,
-      response,
-    };
-  }
-
-  getOrCreateTheme(params: { title: string }) {
-    return {
-      id: this.getTheme(params)?.id ?? this.createTheme(params).id,
-    };
-  }
-
-  deleteTheme({ themeId }: { themeId: string }) {
-    return this.client.delete(`/api/themes/${themeId}`);
-  }
-
-  getTopic({
-    themeId,
-    title,
-  }: {
-    themeId: string;
-    title: string;
-  }): Topic | undefined {
-    return this.getThemes()
-      .flatMap(theme => theme.topics)
-      .find(topic => topic.themeId === themeId && topic.title === title);
-  }
-
-  createTopic({ themeId, title }: { themeId: string; title: string }) {
-    const { response, json } = this.client.post<{ id: string }>(
-      `/api/topics`,
-      JSON.stringify({
-        themeId,
-        title,
-      }),
-      applicationJsonHeaders,
-    );
-
-    console.log(`Created Topic ${title}`);
-
-    return {
-      id: json.id,
-      response,
-    };
-  }
-
-  getOrCreateTopic(params: { themeId: string; title: string }) {
-    return {
-      id: this.getTopic(params)?.id ?? this.createTopic(params).id,
-    };
-  }
-
-  deleteTopic({ topicId }: { topicId: string }) {
-    return this.client.delete(`/api/topics/${topicId}`);
-  }
-
-  getPublications({ topicId }: { topicId: string }): PublicationAndReleases[] {
-    // TODO: update this to use the new endpoint once the 'me' endpoint is removed
-    // https://dfedigital.atlassian.net/browse/EES-3577
-    const { json } = this.client.get<PublicationAndReleases[]>(
-      `/api/me/publications?topicId=${encodeURI(topicId)}`,
-      applicationJsonHeaders,
-    );
-    return json;
-  }
-
-  getPublication({
-    topicId,
-    title,
-  }: {
-    topicId: string;
-    title: string;
-  }): PublicationAndReleases | undefined {
-    return this.getPublications({ topicId }).find(
-      publication =>
-        publication.topicId === topicId && publication.title === title,
-    );
-  }
-
-  createPublication({ topicId, title }: { topicId: string; title: string }) {
-    const { response, json } = this.client.post<{ id: string }>(
-      `/api/publications`,
-      JSON.stringify({
-        topicId,
-        title,
-        contact: {
-          contactName: 'Team Contact',
-          contactTelNo: '12345',
-          teamEmail: 'team@example.com',
-          teamName: 'Team',
-        },
-        summary: `${title} summary`,
-      }),
-      applicationJsonHeaders,
-    );
-
-    console.log(`Created Publication ${title}`);
-
-    return {
-      id: json.id,
-      response,
-    };
-  }
-
-  getOrCreatePublication(params: { topicId: string; title: string }) {
-    return {
-      id: this.getPublication(params)?.id ?? this.createPublication(params).id,
-    };
-  }
-
-  getRelease({
-    topicId,
-    publicationTitle,
-    year,
-  }: {
-    topicId: string;
-    publicationTitle: string;
-    year: number;
-  }): Release | undefined {
-    const publication = this.getPublication({
-      topicId,
-      title: publicationTitle,
-    });
-    return publication?.releases.find(release => release.year === year);
-  }
-
-  createRelease({
-    publicationId,
-    year,
-    timePeriodCoverage,
-  }: {
-    publicationId: string;
-    year: number;
-    timePeriodCoverage: 'AY' | 'FY';
-  }) {
-    const { response, json } = this.client.post<{ id: string }>(
-      `/api/publications/${publicationId}/releases`,
-      JSON.stringify({
-        publicationId,
-        year,
-        timePeriodCoverage: {
-          value: timePeriodCoverage,
-        },
-        type: 'NationalStatistics',
-      }),
-      applicationJsonHeaders,
-    );
-
-    console.log(`Created Release ${year}`);
-
-    return {
-      id: json.id,
-      response,
-    };
-  }
-
-  getOrCreateRelease({
-    topicId,
-    publicationId,
-    publicationTitle,
-    year,
-    timePeriodCoverage,
-  }: {
-    topicId: string;
-    publicationId: string;
-    publicationTitle: string;
-    year: number;
-    timePeriodCoverage: 'AY' | 'FY';
-  }) {
-    return {
-      id:
-        this.getRelease({ topicId, publicationTitle, year })?.id ??
-        this.createRelease({ publicationId, year, timePeriodCoverage })?.id,
-    };
-  }
-
-  getDataFile({
-    releaseId,
-    dataFileName,
-  }: {
-    releaseId: string;
-    dataFileName: string;
-  }) {
-    const { json: releaseFiles } = this.client.get<
-      { id: string; fileName: string }[]
-    >(`/api/release/${releaseId}/data`, applicationJsonHeaders);
-    return releaseFiles.find(file => file.fileName === dataFileName);
-  }
-
-  uploadDataFile({
-    title,
-    releaseId,
-    dataFile,
-    metaFile,
-  }: {
-    title: string;
-    releaseId: string;
-    dataFile: {
-      file: ArrayBuffer;
-      filename: string;
-    };
-    metaFile: {
-      file: ArrayBuffer;
-      filename: string;
-    };
-  }) {
-    const uploadBody = {
-      title,
-      file: http.file(dataFile.file, dataFile.filename),
-      metaFile: http.file(metaFile.file, metaFile.filename),
-    };
-
-    const { response, json } = this.client.post<{ id: string }>(
-      `/api/release/${releaseId}/data?title=${encodeURI(title)}`,
-      uploadBody,
-    );
-
-    const { id: fileId } = json;
-
-    return {
-      id: fileId,
-      response,
-    };
-  }
-
-  waitForDataFileToImport({
-    releaseId,
-    fileId,
-    pollingDelaySeconds = 5,
-    onStatusCheckFailed,
-    onStatusReceived,
-    onImportFailed,
-    onImportCompleted,
-    onImportExceededTimeout,
-  }: {
-    releaseId: string;
-    fileId: string;
-    pollingDelaySeconds?: number;
-    onStatusCheckFailed?: (statusResponse: RefinedResponse<'text'>) => void;
-    onStatusReceived?: (importStatus: string) => void;
-    onImportFailed?: (importStatus: string) => void;
-    onImportCompleted?: () => void;
-    onImportExceededTimeout?: () => void;
-  }) {
-    const importStartTime = Date.now();
-    const importExpireTime = importStartTime + TestData.maxImportWaitTimeMillis;
-
-    while (Date.now() < importExpireTime) {
-      const { importStatus, response } = this.getImportStatus({
-        releaseId,
-        fileId,
-      });
-
-      if (response.status !== 200 && onStatusCheckFailed) {
-        onStatusCheckFailed(response);
-      } else {
-        if (onStatusReceived) {
-          onStatusReceived(importStatus);
-        }
-
-        if (['FAILED', 'CANCELLED'].includes(importStatus)) {
-          if (onImportFailed) {
-            onImportFailed(importStatus);
-          }
-          return;
-        }
-
-        if (['COMPLETE'].includes(importStatus)) {
-          if (onImportCompleted) {
-            onImportCompleted();
-          }
-          return;
-        }
-      }
-
-      sleep(pollingDelaySeconds);
-    }
-
-    if (onImportExceededTimeout) {
-      onImportExceededTimeout();
-    }
-  }
-
-  getOrImportDataFile({
-    title,
-    releaseId,
-    dataFile,
-    metaFile,
-  }: {
-    title: string;
-    releaseId: string;
-    dataFile: {
-      file: ArrayBuffer;
-      filename: string;
-    };
-    metaFile: {
-      file: ArrayBuffer;
-      filename: string;
-    };
-  }) {
-    return {
-      id:
-        this.getDataFile({ releaseId, dataFileName: dataFile.filename })?.id ??
-        this.uploadDataFile({ title, releaseId, dataFile, metaFile })?.id,
-    };
-  }
-
-  getImportStatus({
-    releaseId,
-    fileId,
-  }: {
-    releaseId: string;
-    fileId: string;
-  }) {
-    const { response, json } = this.client.get<{ status: string }>(
-      `/api/release/${releaseId}/data/${fileId}/import/status`,
-    );
-    return {
-      importStatus: json.status,
-      response,
-    };
-  }
-
-  getSubjects({ releaseId }: { releaseId: string }) {
-    const { response, json } = this.client.get<{ id: string; name: string }[]>(
-      `/api/data/releases/${releaseId}/subjects`,
-    );
-    return {
-      subjects: json.map(subject => ({
-        id: subject.id,
-        name: subject.name,
-      })),
-      response,
-    };
   }
 
   getSubjectMeta({
@@ -459,7 +56,7 @@ class DataService {
     subjectId: string;
   }) {
     const { response, json } = this.client.get<SubjectMeta>(
-      `/api/data/release/${releaseId}/meta/subject/${subjectId}`,
+      `/data/release/${releaseId}/meta/subject/${subjectId}`,
     );
     return {
       subjectMeta: json,
@@ -468,7 +65,7 @@ class DataService {
   }
 
   tableQuery({
-    releaseId,
+    publicationId,
     subjectId,
     filterIds,
     indicatorIds,
@@ -478,7 +75,7 @@ class DataService {
     endYear,
     endCode,
   }: {
-    releaseId: string;
+    publicationId: string;
     subjectId: string;
     filterIds: string[];
     indicatorIds: string[];
@@ -489,13 +86,14 @@ class DataService {
     endCode: string;
   }) {
     const { response, json } = this.client.post<{ results: { id: string }[] }>(
-      `/api/data/tablebuilder/release/${releaseId}`,
+      '/tablebuilder',
       JSON.stringify({
         filters: filterIds,
         includeGeoJson: false,
         indicators: indicatorIds,
         locationIds,
         subjectId,
+        publicationId,
         timePeriod: {
           startYear,
           startCode,
@@ -510,136 +108,11 @@ class DataService {
       response,
     };
   }
-
-  addDataGuidance({
-    releaseId,
-    content = '<p>Test Content</p>',
-    subjects,
-  }: {
-    releaseId: string;
-    content?: string;
-    subjects: {
-      id: string;
-      content: string;
-    }[];
-  }) {
-    const { response } = this.client.patch(
-      `/api/release/${releaseId}/data-guidance`,
-      JSON.stringify({
-        content,
-        subjects,
-      }),
-      applicationJsonHeaders,
-    );
-    return {
-      response,
-    };
-  }
-
-  approveRelease({
-    releaseId,
-    notifySubscribers = false,
-    latestInternalReleaseNote = 'Approved',
-  }: {
-    releaseId: string;
-    notifySubscribers?: boolean;
-    latestInternalReleaseNote?: string;
-  }) {
-    const { response } = this.client.post(
-      `/api/releases/${releaseId}/status`,
-      JSON.stringify({
-        notifySubscribers,
-        latestInternalReleaseNote,
-        approvalStatus: 'Approved',
-        publishMethod: 'Immediate',
-      }),
-      applicationJsonHeaders,
-    );
-    return {
-      response,
-    };
-  }
-
-  getReleaseApprovalStatus({ releaseId }: { releaseId: string }) {
-    const { response, json } = this.client.get<{ overallStage: OverallStage }>(
-      `/api/releases/${releaseId}/stage-status`,
-      applicationJsonHeaders,
-      [204],
-    );
-    return {
-      status: json.overallStage,
-      response,
-    };
-  }
-
-  waitForReleaseToBePublished({
-    releaseId,
-    pollingDelaySeconds = 5,
-    onStatusCheckFailed,
-    onStatusReceived,
-    onPublishingFailed,
-    onPublishingCompleted,
-    onPublishingExceededTimeout,
-  }: {
-    releaseId: string;
-    pollingDelaySeconds?: number;
-    onStatusCheckFailed?: (statusResponse: RefinedResponse<'text'>) => void;
-    onStatusReceived?: (status: string) => void;
-    onPublishingFailed?: (status: string) => void;
-    onPublishingCompleted?: () => void;
-    onPublishingExceededTimeout?: () => void;
-  }) {
-    const publishingStartTime = Date.now();
-    const publishingExpireTime =
-      publishingStartTime + TestData.maxPublishingWaitTimeMillis;
-
-    while (Date.now() < publishingExpireTime) {
-      const { status, response } = this.getReleaseApprovalStatus({ releaseId });
-
-      if (response.status === 204) {
-        sleep(pollingDelaySeconds);
-        return;
-      }
-
-      console.log(`.... ${status}`);
-
-      if (response.status !== 200 && onStatusCheckFailed) {
-        onStatusCheckFailed(response);
-      } else {
-        if (onStatusReceived) {
-          onStatusReceived(status);
-        }
-
-        const failureStates: OverallStage[] = ['Failed', 'Invalid'];
-
-        if (failureStates.includes(status)) {
-          if (onPublishingFailed) {
-            onPublishingFailed(status);
-          }
-          return;
-        }
-
-        if (status === 'Complete') {
-          if (onPublishingCompleted) {
-            onPublishingCompleted();
-          }
-          return;
-        }
-      }
-
-      sleep(pollingDelaySeconds);
-    }
-
-    if (onPublishingExceededTimeout) {
-      onPublishingExceededTimeout();
-    }
-  }
 }
 
 export default function createDataService(
   adminUrl: string,
-  accessToken: string,
   checkResponseStatus = true,
 ) {
-  return new DataService(adminUrl, accessToken, checkResponseStatus);
+  return new DataService(adminUrl, checkResponseStatus);
 }
